Update rate_of_return field to DecimalField in InformationPlan; enhance payment ID validation in PaymentDocument; improve error handling for amount input in TransmissionViewset.

+++ plan/models.py
@@ -214,7 +214,7 @@ class Plans (models.Model):
 
 class InformationPlan (models.Model):
     plan = models.ForeignKey(Plan, on_delete = models.CASCADE)
-    rate_of_return = models.BigIntegerField (null=True , blank= True)
+    rate_of_return = models.DecimalField(max_digits=10, decimal_places=2, null=True, blank=True)
     status_second_option = [
          ('1','1'),
          ('2','2'),
